🌟 add charts

+++ src/main/kotlin/com/teamxenox/codoc19/data/entities/Chart.kt
@@ -1,6 +1,5 @@
 package com.teamxenox.codoc19.data.entities
 
-import java.util.*
 import javax.persistence.*
 
 @Entity(name = "charts")
@@ -8,7 +7,7 @@ class Chart {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Long? = null
+    var id: Int? = null
     var userId: Int? = null
     lateinit var tgFileId: String
